change data format from double to single of the raw data when imporing h5 data

+++ @UserCallback/DataSheetSelect.m
@@ -1,5 +1,4 @@
 function DataSheetSelect(app,event)
-
 if event.Source == app.CurrentDataSheetSpinner
     SourceItem = 'Spinner';
 elseif event.Source == app.EigerMasterMenu
@@ -17,10 +16,9 @@ switch SourceItem
         app.CurrentData.RequestSN = app.CurrentDataSheetSpinner.Value;
 end
 
-app.CurrentData.RawData = double(EigerDataFunc.ReadEigerHDF5Data(app.CurrentData.MasterInfo,app.CurrentData.RequestSN,[],[]));
+app.CurrentData.RawData = single(EigerDataFunc.ReadEigerHDF5Data(app.CurrentData.MasterInfo,app.CurrentData.RequestSN,[],[]));
 [~,Title,~] = fileparts(app.CurrentData.MasterInfo.MasterFP); Title = strrep(Title,'_master','');
 app.CurrentData.Title = sprintf('%s#%d',Title,app.CurrentData.RequestSN);
-
 
 GeneralFunc.DataProcess(app,event);
 GeneralFunc.PlotCurrentProfile(app,event);
